requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
psycopg2
python-dotenv
whitenoise
azure-identity